requirements.txt
sounddevice==0.5.2 # 播放音訊
librosa==0.11.0 # 讀取/寫入音訊檔案
numpy==2.2.6
scipy==1.15.3
faster-whisper==1.1.1
transformers==4.54.1
langchain==0.3.27
langchain-huggingface==0.3.1
langchain-openai==0.3.33
python-dotenv==1.1.1
torch==2.7.1
accelerate==1.9.0
jinja2==3.1.6
deepgram-sdk==4.8.1
